pyproject.toml
[tool.poetry]
name = "neuro-bot"
version = "0.1.0"
description = ""
authors = ["Nikita Titarenko <ntv.cnd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
bs4 = "^0.0.2"
lxml = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
